Don't run shutdown at SIGTERM
Don't command the MCU to shut down the system at SIGTERM. The daemon can't
make a difference between a "normal" shutdown and a reboot. If the dns320l-mcu
process is killed as part of a reboot, or a simple daemon restart, the MCU
will be stuck in shutdown mode, and cuts the next bootup in half.