Use /sys/class/gpio/mcu_power for monitoring the MCU power line
It would be much nicer to have this as a configurable value, but for now
(and ad inifinitum), let's hardcode this, and depend on being it exported
from DTS.

Bottom line is, the /sys/class/gpio/mcu_power/value is 1 at default,
and is pulled down (set to 0) by the MCU, if the power button is
pressed for 5 seconds, i.e. user initiates a shutdown. The mcu daemon
runs "halt", then sends a DeviceShutDown with a (configurable) 30 second
timeout to the MCU, to cut the power after this interval.